Extract a superclass
