+++ hapi-fhir-jpaserver-base/src/main/java/ca/uhn/fhir/jpa/bulk/job/BulkItemReader.java
@@ -20,15 +20,12 @@ package ca.uhn.fhir.jpa.bulk.job;
  * #L%
  */
 
-import ca.uhn.fhir.context.FhirContext;
 import ca.uhn.fhir.context.RuntimeResourceDefinition;
 import ca.uhn.fhir.interceptor.model.RequestPartitionId;
-import ca.uhn.fhir.jpa.api.dao.DaoRegistry;
 import ca.uhn.fhir.jpa.api.dao.IFhirResourceDao;
 import ca.uhn.fhir.jpa.batch.log.Logs;
 import ca.uhn.fhir.jpa.dao.IResultIterator;
 import ca.uhn.fhir.jpa.dao.ISearchBuilder;
-import ca.uhn.fhir.jpa.dao.SearchBuilderFactory;
 import ca.uhn.fhir.jpa.dao.data.IBulkExportJobDao;
 import ca.uhn.fhir.jpa.entity.BulkExportJobEntity;
 import ca.uhn.fhir.jpa.model.search.SearchRuntimeDetails;
@@ -38,11 +35,8 @@ import ca.uhn.fhir.jpa.searchparam.SearchParameterMap;
 import ca.uhn.fhir.rest.api.server.storage.ResourcePersistentId;
 import ca.uhn.fhir.rest.param.DateRangeParam;
 import ca.uhn.fhir.util.UrlUtil;
-import org.hl7.fhir.instance.model.api.IBaseResource;
 import org.slf4j.Logger;
-import org.springframework.batch.item.ItemReader;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.beans.factory.annotation.Value;
 
 import java.util.ArrayList;
 import java.util.Arrays;
@@ -51,89 +45,28 @@ import java.util.List;
 import java.util.Map;
 import java.util.Optional;
 
-public class BulkItemReader implements ItemReader<List<ResourcePersistentId>> {
+/**
+ * Basic Bulk Export implementation which simply reads all type filters and applies them, along with the _since param
+ * on a given resource type.
+ */
+public class BulkItemReader extends BaseBulkItemReader {
 	private static final Logger ourLog = Logs.getBatchTroubleshootingLog();
-	Iterator<ResourcePersistentId> myPidIterator;
 
 
-	@Value("#{jobParameters['" + BulkExportJobConfig.READ_CHUNK_PARAMETER + "']}")
-	private Long myReadChunkSize;
-	@Value("#{jobExecutionContext['"+ BulkExportJobConfig.JOB_UUID_PARAMETER+"']}")
-	private String myJobUUID;
-	@Value("#{stepExecutionContext['resourceType']}")
-	private String myResourceType;
-
-	@Autowired
-	private IBulkExportJobDao myBulkExportJobDao;
-	@Autowired
-	private DaoRegistry myDaoRegistry;
-	@Autowired
-	private FhirContext myContext;
-	@Autowired
-	private SearchBuilderFactory mySearchBuilderFactory;
-
-	@Autowired
-	private MatchUrlService myMatchUrlService;
-
-	private void loadResourcePids() {
-		Optional<BulkExportJobEntity> jobOpt = myBulkExportJobDao.findByJobId(myJobUUID);
-		if (!jobOpt.isPresent()) {
-			ourLog.warn("Job appears to be deleted");
-			return;
-		}
-		BulkExportJobEntity jobEntity = jobOpt.get();
-		ourLog.info("Bulk export starting generation for batch export job: {}", jobEntity);
-
-		IFhirResourceDao<?> dao = myDaoRegistry.getResourceDao(myResourceType);
-
+	@Override
+	Iterator<ResourcePersistentId> getResourcePidIterator() {
 		ourLog.info("Bulk export assembling export of type {} for job {}", myResourceType, myJobUUID);
 
-		RuntimeResourceDefinition def = myContext.getResourceDefinition(myResourceType);
-		Class<? extends IBaseResource> nextTypeClass = def.getImplementingClass();
-		ISearchBuilder sb = mySearchBuilderFactory.newSearchBuilder(dao, myResourceType, nextTypeClass);
 
-		SearchParameterMap map = createSearchParameterMapFromTypeFilter(jobEntity, def);
-
-		if (jobEntity.getSince() != null) {
-			map.setLastUpdated(new DateRangeParam(jobEntity.getSince(), null));
-		}
-
-		map.setLoadSynchronous(true);
+		SearchParameterMap map = createSearchParameterMapForJob();
+		ISearchBuilder sb = getSearchBuilderForLocalResourceType();
 		IResultIterator myResultIterator = sb.createQuery(map, new SearchRuntimeDetails(null, myJobUUID), null, RequestPartitionId.allPartitions());
+
 		List<ResourcePersistentId> myReadPids = new ArrayList<>();
 		while (myResultIterator.hasNext()) {
 			myReadPids.add(myResultIterator.next());
 		}
-		myPidIterator = myReadPids.iterator();
+		return myReadPids.iterator();
 	}
 
-	private SearchParameterMap createSearchParameterMapFromTypeFilter(BulkExportJobEntity theJobEntity, RuntimeResourceDefinition theDef) {
-		SearchParameterMap map = new SearchParameterMap();
-		Map<String, String[]> requestUrl = UrlUtil.parseQueryStrings(theJobEntity.getRequest());
-		String[] typeFilters = requestUrl.get(JpaConstants.PARAM_EXPORT_TYPE_FILTER);
-		if (typeFilters != null) {
-			Optional<String> filter = Arrays.stream(typeFilters).filter(t -> t.startsWith(myResourceType + "?")).findFirst();
-			if (filter.isPresent()) {
-				String matchUrl = filter.get();
-				map = myMatchUrlService.translateMatchUrl(matchUrl, theDef);
-			}
-		}
-		return map;
-	}
-
-	@Override
-	public List<ResourcePersistentId> read() {
-		if (myPidIterator == null) {
-			loadResourcePids();
-		}
-		int count = 0;
-		List<ResourcePersistentId> outgoing = new ArrayList<>();
-		while (myPidIterator.hasNext() && count < myReadChunkSize) {
-			outgoing.add(myPidIterator.next());
-			count += 1;
-		}
-
-		return outgoing.size() == 0 ? null : outgoing;
-
-	}
 }
